This seems to be a working version of the software.
After some basic testing, this looks like a version of the software that works.
This commit also includes some ancillary programs that help set up a testing
environment and help with testing in general, but they are just for helping, not
even close to being okay to use for production.

Although this seems to work okay, for now it looks like this doesn't handle
messages that have lines longer than about 1015 characters (not including line
endings). That would be one of the next steps for the program. Also adding
getopt functionality to the pop.c program and daemoninzing the whole thing.
Lastly, adding a user for authpop.c and pledging all three programs would
probably finish it off.

For more general improvements, we could add some (D)DOS protection by
remembering IP addresses that fail to log in and whatnot.